backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0

# OCR & Image Processing
pytesseract==0.3.10
Pillow>=9.5.0,<11
pdf2image==1.16.3

# LLM / AI
openai>=1.12.0,<2.0
python-dotenv==1.0.0

# Vector DB (RAG)
chromadb==0.4.15

# Database
mysql-connector-python==8.2.0

# Production server
gunicorn==21.2.0

# Frontend
streamlit
plotly
